#49 Added documentation.
