+++ src/main/java/org/ahocorasick/trie/Payload.java
@@ -1,5 +1,12 @@
 package org.ahocorasick.trie;
 
+/**
+ * Payload holds the matched keyword and some payload-data.
+ * 
+ * @author Daniel Beck
+ *
+ * @param <T> The type of the wrapped payload data.
+ */
 public class Payload<T> implements Comparable<Payload<T>> {
 
     private final String keyword;
