+++ src/main/java/org/ahocorasick/trie/PayloadMatchToken.java
@@ -1,5 +1,15 @@
 package org.ahocorasick.trie;
 
+/**
+ * PayloadMatchToken holds a text ("the fragment") an emits some output.
+ * <p>
+ * It matches a search term - so its <code>isMatch</code>-method returns always
+ * true..
+ * 
+ * @author Daniel Beck
+ *
+ * @param <T> The Type of the emitted payloads.
+ */
 public class PayloadMatchToken<T> extends PayloadToken<T> {
 
     private final PayloadEmit<T> emit;
