+++ src/main/java/org/ahocorasick/trie/handler/PayloadEmitDelegateHandler.java
@@ -1,27 +1,18 @@
 package org.ahocorasick.trie.handler;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
 import org.ahocorasick.trie.Emit;
 import org.ahocorasick.trie.PayloadEmit;
 
-public class PayloadEmitDelegateHandler implements StatefulPayloadEmitHandler<String> {
+/**
+ * Convenience wrapper class that delegates every method to a EmitHandler.
+ */
+public class PayloadEmitDelegateHandler implements PayloadEmitHandler<String> {
 
-    private StatefulEmitHandler handler;
+    private EmitHandler handler;
 
-    public PayloadEmitDelegateHandler(StatefulEmitHandler handler) {
+    public PayloadEmitDelegateHandler(EmitHandler handler) {
         this.handler = handler;
 
-    }
-
-    private static List<PayloadEmit<String>> asEmits(Collection<Emit> emits) {
-        List<PayloadEmit<String>> result = new ArrayList<>();
-        for (Emit emit : emits) {
-            result.add(new PayloadEmit<String>(emit.getStart(), emit.getEnd(), emit.getKeyword(), null));
-        }
-        return result;
     }
 
     @Override
@@ -30,9 +21,4 @@ public class PayloadEmitDelegateHandler implements StatefulPayloadEmitHandler<St
         return handler.emit(newEmit);
     }
 
-    @Override
-    public List<PayloadEmit<String>> getEmits() {
-        List<Emit> emits = this.handler.getEmits();
-        return asEmits(emits);
-    }
 }
